+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to Recall will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [1.2.1](https://github.com/fnune/recall.nvim/releases/tag/1.2.1) - 2026-03-09
+
+### Added
+
+- Navigation now searches across all tab pages when jumping to a mark, switching to the correct
+  tab page if the target buffer is already displayed there.
 
 ### Fixed